Arrumando botão Delete

+++ controllers/MarcasController.js
@@ -41,24 +41,24 @@ router.get("/admin/marcas", (req, res) => {
 
 router.post("/marcas/delete", (req, res) => {
     var id = req.body.id;
-    if(id != undefined){
-        if(isNaN(id)){
-            //deleteando caso o id seja igual o id recebido
-            Marca.destroy({
-                where: {
-                    id: id
-                }
-            }).then(() => {
-                res.redirect("/admin/marcas");
-            });
-        }else{
+    console.log("ID recebido para deletar:", id);  // Verifique o valor aqui
+    if (id != undefined && !isNaN(id)) {
+        Marca.destroy({
+            where: {
+                id: id
+            }
+        }).then(() => {
+            console.log("Marca deletada com sucesso.");
             res.redirect("/admin/marcas");
-        }
-    }else{
+        }).catch((error) => {
+            console.log("Erro ao deletar marca:", error);
+            res.redirect("/admin/marcas");
+        });
+    } else {
         res.redirect("/admin/marcas");
     }
-    
 });
+
 
 
 module.exports = router;
